Move base entities to their own folder/namespace
