+++ Application/Dtos/GatewayDto.cs
@@ -1,5 +1,6 @@
 ﻿using Application.DataAnnotations;
 using Domain.Entities;
+using Domain.Entities.Base;
 using Domain.Localization;
 using System;
 using System.Collections.Generic;
