+++ Domain/Entities/Base/BaseGenericEntity.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Entities
+namespace Domain.Entities.Base
 {
     public abstract class BaseGenericEntity<TKey>
     {
