+++ Domain/Entities/Gateway.cs
@@ -1,4 +1,5 @@
 ﻿using Domain.Entities;
+using Domain.Entities.Base;
 using System.Collections.Generic;
 
 namespace Domain.Entities
