+++ Domain/Entities/Provider.cs
@@ -1,4 +1,5 @@
 ﻿using Domain.Entities;
+using Domain.Entities.Base;
 using System;
 using System.Collections.Generic;
 using System.Linq;
