+++ Domain/IRepositories/IRepository.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
+using Domain.Entities.Base;
 
 namespace Domain.IRepositories
 {
