+++ Domain/Specification/Specification.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using AutoMapper.Extensions.ExpressionMapping;
 using Domain.Entities;
+using Domain.Entities.Base;
 using System;
 using System.Collections.Generic;
 using System.Linq;
