+++ Infraestructure/Domain/UnitOfWork/UnitOfWorkContainer.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
+using Domain.Entities.Base;
 
 namespace Infraestructure.Domain.UnitOfWork
 {
